Add Compose-based Celery project with list_collection task
This commit adds a Docker Compose project which brings up an iRODS
server with its catalog database (based on the irods_demo project),
a Redis service, and a Celery project.

This commit adds a Celery task called list_collection that lists the
data objects and collections inside the collection specified by
logical_path. This will be the basis upon which parallel removal
will be built. There is an unimplemented recursive option as well.